The example import does not work on MAC
Importing from pansharpen did not work for me when installing via pip on a Mac. I had to do what I am proposing in the documentation change (use rio_pansharpen) instead to get it working. Not sure if it's a difference in python versions or a new change that is not reflected on the docs.